booking_ai_agent/requirements.txt
flask
groq
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv


gunicorn


Flask
gunicorn
requests
python-dotenv
groq
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0